mcp: carve tool_gate.rs (pure move)
The tool-surface gates: rewind-only/managed-context/fission tool sets,
per-profile advertisement filtering, default-session-id injection, the
IAM operation map (mcp_tool_operation), and the manual HTTP tool
definitions — 406 body lines and 3 tests verbatim. Usual glue;
purity-checked; mcp:: tests 178/178.
